Add files via upload
Gaussian Mixture Model (GMM) on MNIST
This code applies GMM to the MNIST dataset, modeling the data as a mixture of several Gaussian distributions. GMM is useful for soft clustering, where each data point is assigned a probability of belonging to multiple clusters. The code also includes dimensionality reduction using PCA and visualizes the resulting clusters in 2D.