Add files via upload
PCA (Principal Component Analysis) on MNIST
This code applies PCA to the MNIST dataset for dimensionality reduction. PCA identifies the most important features that capture the majority of the data's variance. The project includes visualizations of the data projected in 2D and 3D, along with an evaluation of how many principal components are needed to retain 95% variance.