Add files via upload
LDA (Latent Dirichlet Allocation) on MNIST (In Progress)
This project applies LDA for topic modeling on the MNIST dataset, with PCA used to reduce dimensionality and speed up the process. The goal is to explore how LDA can uncover hidden structures in digit data by treating each digit as a "document" of pixel features. The project is currently under development, and further optimizations are in progress.