Add files via upload
SOM (Self-Organizing Maps) on MNIST (In Progress)
This project applies SOM to the MNIST dataset for unsupervised clustering of handwritten digits. The goal is to visualize how the SOM algorithm organizes digits into clusters based on their pixel features. The project is currently under development, with ongoing efforts to optimize training and improve the visualization of clusters.