Add files via upload
Mean Shift Clustering on MNIST (In Progress)
This project applies Mean Shift Clustering to the MNIST dataset for unsupervised classification of handwritten digits. Mean Shift is a non-parametric clustering technique that does not require the number of clusters to be specified in advance. PCA is used for dimensionality reduction to make the clustering process more efficient. The project is currently under development, with optimizations and further analysis in progress.