Add files via upload
Spectral Clustering on MNIST (In Progress)
This project applies Spectral Clustering to the MNIST dataset for unsupervised classification of handwritten digits. Spectral Clustering uses graph-based techniques to partition the data into clusters based on feature similarity. PCA is used for dimensionality reduction to enhance the efficiency of the clustering process. The project is currently under development, with ongoing analysis of clustering accuracy and visualization.